Version 0.2.2: bugfixes and performance improvements.
Fixes panic when attempting to `rotate_left` or `rotate_right` an empty
`DigitsIterator`.

Add examples to the documentation of `DigitsIterator`.

Performance improvement when initializing a `DigitsIterator` with the
`new` method.

Performance improvement when using the `reverse_digits` method of the
`IntoDigits` trait by avoiding the costly initialization of a
`DigitsIterator` struct.