IntoDigits: use default permutation implementation for specific ones.
Use `IntoDigits::is_permutation` method for the default implementations
of `is_decimal_permutation` and `is_binary_permutation`.

This has a slightly negative impact on performance. While this is not
yet remedied for decimals, the default implementation is replaced for
`is_binary_permutation` to leverage the `count_ones` method for the
standard unsigned types, yielding a massive improvement.